refactor: move lottie splash screen logic into hook
