+++ src/components/LottieSplashScreen.tsx
@@ -1,60 +1,36 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { Modal, Animated, Platform } from 'react-native';
+import { FunctionComponent } from 'react';
+import { Modal, Animated, useColorScheme } from 'react-native';
 import Lottie from 'lottie-react-native';
 
 import lottieAnimation from 'src/assets/splash/lottie_animated_logo.json';
+import {
+  UseLottieSplashScreenParams,
+  useLottieSplashScreen,
+} from 'src/hooks/useLottieSplashScreen';
+import { colorSchemePrimaryBgMap } from '_utils/colorSchemePrimaryBgMap';
+import { StatusBar } from 'expo-status-bar';
 
-interface SplashScreenProps {
-  animationFadeOut: boolean;
-  onHidden?: () => void;
-}
-
-const FADE_IN_DURATION = 1000;
-const FADE_OUT_DURATION = 500;
-
-const SplashScreen: React.FunctionComponent<SplashScreenProps> = ({
+const SplashScreen: FunctionComponent<UseLottieSplashScreenParams> = ({
   animationFadeOut,
   onHidden,
 }) => {
-  const [visible, setVisible] = useState(true);
-  const fadeAnim = useRef(new Animated.Value(0)).current;
-
-  const fadeIn = () => {
-    Animated.timing(fadeAnim, {
-      toValue: 1,
-      duration: FADE_IN_DURATION,
-      useNativeDriver: true,
-    }).start();
-  };
-
-  const fadeOut = () => {
-    Animated.timing(fadeAnim, {
-      toValue: 0,
-      duration: FADE_OUT_DURATION,
-      useNativeDriver: true,
-    }).start(() => {
-      setVisible(false);
-      if (onHidden) onHidden();
+  const { visible, fadeAnim, onLottieAnimationComplete } =
+    useLottieSplashScreen({
+      animationFadeOut,
+      onHidden,
     });
-  };
 
-  useEffect(() => {
-    fadeIn();
-
-    if (animationFadeOut) {
-      fadeOut();
-    }
-  }, [animationFadeOut]);
-
-  const onLottieAnimationComplete = () => {
-    if (animationFadeOut) {
-      fadeOut();
-    }
-  };
+  const colorScheme = useColorScheme();
 
   return (
     <Modal transparent visible={visible}>
-      <Animated.View style={{ flex: 1, opacity: fadeAnim }}>
+      <Animated.View
+        style={{
+          flex: 1,
+          opacity: fadeAnim,
+          backgroundColor: colorSchemePrimaryBgMap[colorScheme ?? 'light'],
+        }}
+      >
         <Lottie
           source={lottieAnimation}
           loop={false}
